尝试修正centos8 8.3.1 下的编译问题。

+++ include/common/s4calendar.h
@@ -17,7 +17,7 @@ public:
 	int getDate(int idx) const
 	{
 		if (idx >= 0) {
-			if (idx >= size()) {
+			if ((size_t)idx >= size()) {
 				ERR("calendar getDate({:d}) ovf!\n", idx);
 				return 20840328;
 			}
@@ -25,7 +25,7 @@ public:
 		}
 		else
 		{
-			if (-idx > size()) {
+			if ((size_t)(-idx) > size()) {
 				ERR("calendar getDate({:d}) - ovf!\n", idx);
 				return 0;
 			}
